pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlms-cosem"
version = "2026.1.0"
description = "A Python library for DLMS/COSEM"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Business Source License 1.1" }
authors = [
  { name = "Henrik Palmlund Wahlgren @ Palmlund Wahlgren Innovative Technology AB", email = "henrik@pwit.se" },
]
maintainers = [
  { name = "Henrik Palmlund Wahlgren @ Palmlund Wahlgren Innovative Technology AB", email = "henrik@pwit.se" },
]
keywords = ["AMR", "Metering", "smart meters", "MDM", "dlms", "cosem"]
classifiers = [
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Communications",
]
dependencies = [
  "attrs>=22.2.0",
  "pyserial>=3.5",
  "cryptography>=35.0.0",
  "asn1crypto>=1.4.0",
  "python-dateutil>=2.8.1",
  "typing-extensions>=3.10",
  "structlog>=22.1.0",
]

[project.optional-dependencies]
docs = ["mkdocs", "mkdocs-material"]
test = ["pytest", "pytest-cov", "pytest-sugar"]

# Key management optional dependencies
keyring = ["keyring"]
yaml = ["pyyaml"]
toml = ["tomli", "tomli-w"]
keys = ["dlms-cosem[keyring,yaml,toml]"]

dev = [
  "pre-commit",
  "mkdocs",
  "mkdocs-material",
  "pytest",
  "pytest-cov",
  "pytest-sugar",
  # Key management dependencies for development
  "pyyaml",
  "tomli",
  "tomli-w",
]

[project.scripts]
dlms-keys = "dlms_cosem.cli.dlms_keys:main"

[project.urls]
Documentation = "https://www.dlms.dev/"
"Bug Tracker" = "https://github.com/u9n/dlms-cosem/issues"
"Source Code" = "https://github.com/u9n/dlms-cosem"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]

[tool.black]
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
